// src/components/common/Extensions/NotificationBanner/NotificationBanner.types.ts
export interface NotificationBannerProps {
  status: StatusType;
  title: string;
  actionText: string;
  actionType: ActionType;
  isFullSize?: boolean;
}

type ActionType = 'redirect' | 'copy-url' | 'call-to-action';

type StatusType = 'success' | 'warning' | 'error';

export interface CopyUrlProps {
  actionText: string;
}